tuanvy updated 16/02/2024

+++ src/main/java/com/shopproject/shopbt/repository/color/ColorRepository.java
@@ -7,5 +7,5 @@ import java.util.Set;
 import java.util.stream.Stream;
 
 public interface ColorRepository extends JpaRepository<Color, Long> {
-    Color findByColorId(Integer id);
+    Color findByColorId(Long id);
 }
